+++ scm-webapp/src/test/java/sonia/scm/net/ahc/DefaultAdvancedHttpClientTest.java
@@ -317,6 +317,28 @@ public class DefaultAdvancedHttpClientTest
     verify(tracer, never()).span(anyString());
   }
 
+  @Test
+  public void shouldNotTraceRequestIfAcceptedResponseCode() throws IOException {
+    when(connection.getResponseCode()).thenReturn(400);
+
+    new AdvancedHttpRequest(client, HttpMethod.GET, "https://www.scm-manager.org").acceptStatusCodes(400).request();
+    verify(tracer).span("HTTP Request");
+    verify(span).label("status", 400);
+    verify(span, never()).failed();
+    verify(span).close();
+  }
+
+  @Test
+  public void shouldTraceRequestAsFailedIfAcceptedResponseCodeDoesntMatch() throws IOException {
+    when(connection.getResponseCode()).thenReturn(401);
+
+    new AdvancedHttpRequest(client, HttpMethod.GET, "https://www.scm-manager.org").acceptStatusCodes(400).request();
+    verify(tracer).span("HTTP Request");
+    verify(span).label("status", 401);
+    verify(span).failed();
+    verify(span).close();
+  }
+
 
   //~--- set methods ----------------------------------------------------------
 
@@ -328,7 +350,7 @@ public class DefaultAdvancedHttpClientTest
   public void setUp()
   {
     configuration = new ScmConfiguration();
-    transformers = new HashSet<ContentTransformer>();
+    transformers = new HashSet<>();
     client = new TestingAdvacedHttpClient(configuration, transformers);
     when(tracer.span(anyString())).thenReturn(span);
   }