+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Delete branches directly in the UI ([#1422](https://github.com/scm-manager/scm-manager/pull/1422))
 - Lookup command which provides further repository information ([#1415](https://github.com/scm-manager/scm-manager/pull/1415))
 - Include messages from scm protocol in modification or merge errors ([#1420](https://github.com/scm-manager/scm-manager/pull/1420))
+- Enhance trace api to accepted status codes ([#1430](https://github.com/scm-manager/scm-manager/pull/1430))
 
 ### Fixed
 - Missing close of hg diff command ([#1417](https://github.com/scm-manager/scm-manager/pull/1417))